restructured-event-page-MVVM
